+++ src/lib.rs
@@ -93,6 +93,7 @@
 
 #![no_std]
 
+#[cfg(feature = "device-selected")]
 use embedded_hal as hal;
 
 #[cfg(feature = "stm32f100")]
@@ -104,26 +105,49 @@ pub use stm32f1::stm32f101 as pac;
 #[cfg(feature = "stm32f103")]
 pub use stm32f1::stm32f103 as pac;
 
+#[cfg(feature = "device-selected")]
 pub use crate::pac as device;
+
+#[cfg(feature = "device-selected")]
 pub use crate::pac as stm32;
 
+#[cfg(feature = "device-selected")]
 pub mod adc;
+#[cfg(feature = "device-selected")]
 pub mod afio;
+#[cfg(feature = "device-selected")]
 pub mod backup_domain;
+#[cfg(feature = "device-selected")]
 pub mod bb;
+#[cfg(feature = "device-selected")]
 pub mod delay;
+#[cfg(feature = "device-selected")]
 pub mod dma;
+#[cfg(feature = "device-selected")]
 pub mod flash;
+#[cfg(feature = "device-selected")]
 pub mod gpio;
+#[cfg(feature = "device-selected")]
 pub mod i2c;
+#[cfg(feature = "device-selected")]
 pub mod prelude;
+#[cfg(feature = "device-selected")]
 pub mod pwm;
+#[cfg(feature = "device-selected")]
 pub mod pwm_input;
+#[cfg(feature = "device-selected")]
 pub mod qei;
+#[cfg(feature = "device-selected")]
 pub mod rcc;
+#[cfg(feature = "device-selected")]
 pub mod rtc;
+#[cfg(feature = "device-selected")]
 pub mod serial;
+#[cfg(feature = "device-selected")]
 pub mod spi;
+#[cfg(feature = "device-selected")]
 pub mod time;
+#[cfg(feature = "device-selected")]
 pub mod timer;
+#[cfg(feature = "device-selected")]
 pub mod watchdog;